requirements.txt
Flask==2.3.2
boto3==1.34.38
python-dotenv==1.0.1